requirements.txt
openai>=1.44.0
requests>=2.32.0
python-dotenv>=1.0.1
